backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
pydantic==2.11.1
python-dotenv==1.0.1
httpx==0.28.1
